Corrected bugs that prevented correct positioning of datapoints (lines, psyms...) with commands PLOTS, XYOUTS and OPLOT when PLOT was done with /[X|Y]LOG.
Added support for NSUM=xxx and /POLAR options used in PLOT, OPLOT.

Treatment for DATAMIN,DATAMAX and CLIP is not OK however. needs some more changes.

Debugged the /NORMAL, /DEVICE and /CONTINUE option of PLOTS. Should be OK even for some (not all) cases when  !P.MULTI is not 0.

Added SIZE=xxx undocumented option as a replacement for CHARSIZE for the XYOUTS command (support for several legacy procedures).