Help_pro refurbishment and code cleanup: part 1
The help routine had been included with the basic_pro set of routines
and it grew to become the largest single routine with the most keywords.
It grew in increments and so is a challenge to decipher.
These changes put much of the formatting code in subroutines called from
help_pro, allowing the flow and logic of help_pro to be more approachable.
Some new features of help will  also be present:
help,/heap
COMMON, OBJECT, METHODS keywords.
Output=<variable> is implemented in many more places.
This is part 1, which is maybe 1/3 of the way finished to help_pro() which is
pretty much the same as it was.

gdlhelp
Greg Jung (aka maynardGK)
gvjung@gmail.com

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# Explicit paths specified without -i or -o; assuming --only paths...
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   src/gdlhelp.cpp
#
# Changes not staged for commit:
#	modified:   CMakeLists.txt
#	modified:   src/devicewin.hpp
#
# Untracked files:
#	dotouch.sh
#	"h\033:"
#	src/gdl
#	testsuite/file_etc/
#	testsuite/launchtest.c
#	touchsrc.sh
#