Implement DELVAR procedure, .RNEW
delvar_pro is found in basic_pro.cpp;
methods to GDL internals are added to facilitate this in envt and in dpro.
The .RNEW command is activated in dinterpreter.cpp
compile option flags STATIC and NOSAVE are in place, although they are not in use.

findvar_pro may be useful in a development context - normally it is not compiled.
DELVAR is to be used only from the $MAIN program, as per IDL docs; if it isn't, an
exception is raised with a message.

Tested since 2016.  Smoke test made for current git build.
tesing procedures requires proper operation of HELP: no output for an undefined
variable is behavior for the new gdlhelp
Stage 1
Greg Jung (aka maynardGK)
gvjung@gmail.com