Help_pro refurbishment and code cleanup: gdlhelp.cpp
The help routine had been included with the basic_pro set of routines
and it grew to become the largest single routine with the most keywords.
It grew in increments and so is a challenge to decipher.
These changes put much of the formatting code in subroutines called from
help_pro, allowing the flow and logic of help_pro to be more approachable.
pretty much the same as it was.

Some new features of help will  also be present:
   Output=<variable> is implemented in many more places.
   names = <reg expression> used with other output to restrict the matches found.
   HEAP numerates object and pointer heap, also more detail if FULL is specified.
   COMMON will list all common blocks anywhere in program
         The method commonPtrs of class DSubUD is added in dpro.hpp
   OBJECT, METHODS keywords for use with objects.
   DEBUG  will set or unset a debug flag that persists for session.

Other behavior changes:
 - For undefined variables, no output is generated unless /FULL keyword is also present.
 - the /STRUCT keyword, which is in effect if a structure is the sole parameter, will
  invoke a recursive description of any sub-structures.  The one-line description
  will always be present, before the detailed structure listing.

testsuite/test_help.pro is a very rudimentary start of several features.

gdlhelp_fin
Greg Jung (maynardGK)
gvjung@gmail.com