basic_fun branch code update.
  These are changes to basic_fun that advance GDL functionality.
 : res=OBJ_HASMETHOD([2 Args])
 : res=ROUTINE_FILEPATH([1 Arg],EITHER,IS_FUNCTION)
Doesn't function ?? but the old code is here waiting to be debugged.
 : res=ARRAY_EQUAL([2 Args],NO_TYPECONV,NOT_EQUAL,QUIET)
ALL keywords now apply (NOT_EQUAL was missing)
i entries in dstruct, datatypes, and in basegdl also had to be made to support this.
 : res=PTR_VALID([1 Arg],CAST,COUNT,GET_HEAP_IDENTIFIER)
 : res=OBJ_VALID([1 Arg],CAST,COUNT,GET_HEAP_IDENTIFIER)
 : res=COMMAND_LINE_ARGS(COUNT,RESET,SET)
  COMMAND_LINE_ARGS is modified to allow insertion of command line arguments while running GDL.
The new options are not IDL compliant and so they are de-activated in libinit, so a developer's
build will be needed to activate it. In order to implement this, the  global vector holding
the args needs to be a string. It is defined in this file and in gdl.cpp:

This file is also revised on branch "heap_refcount" to include that function.

A few fixes and cleanup are also made:
 the defined(__CYGWIN__) macro is quite obsolete and #ifdef _WIN32 is used for
windows compilation.
 PTR_VALID(/cast) is fixed to increment the reference count when a cast is made.
  // for use in COMMAND_LINE_ARGS()
  std::vector<std::string> command_line_args;

 Except for routine_filepath and command_line_args, simple test procedures have been made.