test_ptr_valid.pro using the new heap_gc to test the /CAST keyword
Since this routine calls heap_free and heap_gc to begin actual testing, prior to that
a check is made that there are no pre-existing objects or pointers. This is essentially
a complete basic_fun branch, other mods are in concurrent branches, including
the heap_refcount branch.  Hence, this program avoids the use of heap_refcount.

Greg Jung (aka maynardGK)
gvjung@gmail.com