Patch to avoid accidental assert failures upon keyword lookup.
  To streamline keyword processing we use a trick to look up once and for all the keyword index.
 For almost all keywords, we use
    static keywordnameIx = e->KeywordIx("KEYWORDNAME");
 where the call from the interpreter is given by
	 BaseGDL* function_fun(EnvT* e)  { ...
 or by
                void routine_pro(EnvT *e) { ...
   The indexed list is created in the corresponding call to DLibFun, DLibPro,
DLibFunRetNew etc. as found in the libinit*.cpp files.  The routine EnvT::KeywordIx finds the
index from the provided string.  If there is no corresponding string, pro->FindKey()
returns instead a value of -1.  Through version 0.9.8, the result "-1" is tested with an
assert( val != -1) leading to program exit in an undignified manner.

  This commit is to change that behavior so that, instead of assert() being invoked,
a diagnostic message is issued (3 lines, including info on procedure call and keyword intended).
Then, instead of throwing an exception, the index value is set to zero.  So that accidental
usage of the uninitialized keyword will be interpreted as the first keyword.

Greg Jung (aka maynardGK)
gvjung@gmail.com