bringing the mods from envt_keyword
  file_etc branch is failing on a number of unrelated tests.  Several of these failures
seemed to coincide with a race from envt.cpp line 1160, wheree an assert(val != -1) is coded.
The case where val == -1 is maybe not so uncommon that we should close down the whole circus!
The envt_keyword branch/pull converts this assert to a harmless diagnostic message.
It is included here again, to see if it solves our test failures.