legacy_list: update for LIST,HASH: N_elements(), GDL_CONTAINER
This branch begins the road to the modern versions of list/hash by preserving the old behavior (and actual code implementation) of the pre-existing features, while making the required peripheral modifications to other GDL internals.
The legacy code includes a small bug such that changes to list elements are mostly not possible:
  "alist[k] = findgen(8)" should replace the k'th element of list alist with the right-side expression. Fixing this simple problem in the legacy code would be a) pointless and b) change the apparent baseline code from what is has been for years.

Some of the changes exist in github development separately, awaiting merge/approvals, and are copied to here for use; since they are derived from the same code there should be no conflict issues when merged. In particular, the surgery mentioned in issue #236 - aka SF.net bug 696 (https://sourceforge.net/p/gnudatalanguage/bugs/696/) is made, and it has worked well.  The one adverse side effect required "foreach" coding to notice hashes and lists, and this is also present.

LIST methods move() and swap(), which didn't exist at all, have been copied into the file alongside the legacy code.  Another addition is the "GET" method (list__get) which is not really a list method but intended for the container object; before the GDL_CONTAINER declarations I used those of list to develop this method.  list__get() compiles and exists harmlessly as long as it is not invoked.

Subsequent improvements to list.cpp that are already coded:
   toarray() method is implemented with all features
   Multi-dimensional direct access to list or hash elements (see documention, and some examples in TEST_LIST).
   GDL_CONTAINER is defined, which may lead to IDL_CONTAINER - see in src/pro directory:
;-----------------------------------------
PRO IDL_CONTAINER__Define, class

   Compile_Opt hidden
   class = { IDL_CONTAINER, $
              INHERITS GDL_CONTAINER $
           }

END

Greg Jung (aka maynardGK)
gvjung@gmail.com