legacy_hash: update for HASH - code style update and improvement of existing functions.
  This branch contains functional improvements to the built-in object HASH that are only modestly beyond what has
been available up to gdl 0.9.8. The left-bracket and right-bracket overload routines are functionally the same
although the new coding style has replaced some of that code.

This branch and the "legacy_list" branch begins the road to the modern versions of list/hash by preserving the old behavior
 of the pre-existing features, while making required modifications to other GDL internals.
This branch requires modification only of the hash.cpp routine.
The new LIST/HASH routines will be more co-dependent when the new overload routines are inserted, and this will
be introduced in a single branch, after legacy_list and legacy_hash branched have been merged to master.

The test routine test_hash.pro senses whether it is being run with legacy HASH and will limit expectations accordingly.
Contributions that  exercise more of the existing functions are encouraged, as well as new tests that can only be
 tested in a fully functional IDL environment.

  The method TOSTRUCT() and keywords "EXTRACT, FOLDCASE", "LOWER" should be completely functional now.  A placeholder
for ordered hashes is made: coding such a hash type appears to be a non-trivial project suitable for advanced mathematics.

Greg Jung (aka maynardGK)
gvjung@gmail.com