Begin widget initialization in Objects.cpp
  Objects.cpp is a routine called upon initailization of GDL.  Historically the widget
 initialization was done via a call to GDLWidget::Init(), from this routine.
  Lately it was moved to graphicsdevce.  For whatever reason, The healthy baseline code
prefers it to be in same location.  GDLWidget::Init() is again called and SetWxStarted is
added to the Init() procedure located in gdlwidget.cpp.

   The file testsuite/testfocus.pro is added. it puts out 6 pots with various charsize, thickness
and at different display locations; then various "wshow" are exercised.  In particular,
window 3 and window 0 are iconified and then re-shown.  WShow was re-defined to take
an int instead of bool for the "iconify" flag. Running this exercise deleted these windows
on my (MS win7) machine.  Using the old way of calling (bool) kept these windows from dying.
So an #ifdef _WIN32 block was created around the two methods.

https://github.com/gnudatalanguage/gdl/issues/249

Greg Jung (aka maynardGK)
gvjung@gmail.com