Allow widgets, without widget_draw
In order for a plot window to perform well in a widget group, the device "WX" needs to be utilized.
However, the WX device has a few shortcomings relative to the legacy "deviceX" and "deviceWIN"
 streams used in GDL.  Most notable is the inability to control the keyboard focus of the Wx windows.
config.h.cmake has been modified to include "NO_WIDGET_DRAW".
  If one can forego the draw widget, widgets will co-exist with the standard X and WIN devices.
The build macro NO_WIDGET_DRAW can be used to reject any accidental calls that invoke a draw widget.
Instead, a widget ID of -1 is simply returned with a message, without error.  If this widget ID (-1) is then passed to widget_control, it also will return immediately. This feature is not helpful to those GDL which run with X-windows.
  Additionally, the USE_WINGDI_NOT_WINGCC macro is defined so that, for those with plplot version 13
or later, the "wingdi" device can be used.  It produces marginally better results with text over
use of the "wingcc" device.

Greg Jung (aka maynardGK)
gvjung@gmail.com