Slowly find where OSX/clang test fails.
  This is the baseline build of code for win_plots, where invalid windows plots were
fixed and the keyboard focus was correctly restored for Linux.  The Travis builds
xxx.5, and xxx.10 failed via test_bug_2974380 this is for clang compiled on OSX
systems.

This first commit only submits the repaired devicewin.cpp, .hpp files (which are not used in OSX)
If this passes the next step will be to activate the UnsetFocus() call in WShow.