Improve container routines ADD and GET. test_container.
  Routines to service the method needs of GDL_CONTAINER iare kept in the list.cpp file
because a list is derived from a container.  However a list generates pointers to the data items
it holds, whereas a GDL_CONTAINER holds the pointers and Object IDs directly.
So in list__add, which operates both on list and containers, a "listmode" bool variable is
derived that will distinguish treatments.  Also, since pointers and Object are not to be mixed
in a container, isPtr/!isPtr consitute two other treatments.

  container__get can be called from a list object for the time being; it is not
conforming with IDL but may be interesting in the meantime, while under development.

Already completed, further refinements to LIST will await merge of the pending
branches (legacy_list and legacy_hash) branches into master.  However the container object
 is in  a younger status: this commit revises ::GET substantially (but leaves other routines of list.cpp as-is).
testsuite/test_container.pro is a rudimentary stuff-and-count, beginnings of a good test.

Greg Jung (aka maynardGK)
gvjung@gmail.com