Revise GDL_container::remove method and ::cleanup
  All options of ::remove procedure for GDL_CONTAINER objects are operative.
container__cleanup is fully functional; heap reference decrements are applied
 as are heap reference increments  when a ::GET is used.
Style-motivated revisions made in legacy sections; properties of LISTs are
still maintained to historical norms. The GetSELF routine, replaced by GetOBJ, is
not used but is needed for legacy HASH.

test_container.pro remains the same; It shows that the list objects are not under control.