re-cast test_delvarrnew such that dependent scripts are created in place.
  DELVAR is a routine that should only be callled from $Main level and so
to accomplish that from a subroutine, we shall spawn the GDL process to execute a short script
(i.e. command-line mode, "@") that will in turn, to test "RNEW", run a program.  These scripts
are lost to the normal testing process which occurs in the build area and so we generate them as
needed in the local directory.
  Template files are kept in gdl/testsuite:
	test_delvarrnew.pp.pro
	delvarrnew.pp.pro
	rnewtest.pp.pro
To generate test_delvarrnew.pro, run gdl/testsuite/make_test_delvarrnew.pro

Greg Jung (aka maynardGK)
gvjung@gmail.com