Revise ptr_valid, obj_valid to treat scalar case.
 Previous versions created the shadow long array as
"pL = new DLongGDL( p->Dim() );" which worked but does not act well in the case that a scalar
is used - a scalar was getting created but the internals are not consistent.
So in both cases a separate treatment is used for the scalar case
bool isscalar = p->StrictScalar();
DLongGDL* pL; if(isscalar) pL = new DLongGDL(0); else pL = new DLongGDL( p->Dim());

I expect the tests to pass and so the debug prints are removed.