Combined merger of 5 inter-related branches
  This PR combines five inter-related change branches that have been waiting since before gdl0.9.5
in essentially the same form as now.  Still waiting to mrge, once this has been in master, are the
in-place insertion and extraction of data held in ist and hash containers.
  Because of the extensive editing changes made to LIST and HASH, this intermediate state can still
be termed "legacy" in that the general behavior should be the same except for a few enhancements.
I was able to add the "TOARRAY" method without a lot of trauma.
  Internally, aside from a few routines that service the DELVAR and RNEW procedures, a few small changes
with one exception: the N_Elements() function now always returns the N_Elements() of the GDL variable
referenced to it; LIST and HASH objects do not get an exceptional treatment until the IDL convention
that n_elements( <a list>) will return the count() of that list (or hash).

  The built-in object GDL_CONTAINER is also new, and equivalent to IDL_CONTAINER.  In fact the terms are
equivalenced in the code, rather than IDL_CONTAINER being coded in script; this can be easily undone
 if there are difficulties.

Greg Jung (aka maynardGK)
gvjung@gmail.com