Interim solution with widget_draw option.
  Current state of affairs is that "WX" was supposed to unify plotting and widgets, but
the plplot driver for WX is not quite as capable as a plain X driver or Win driver for the
simple plots.  However incorporating a plot  or image into a widget group requires this. To do so,
set GDL_USE_WX=YES in the environment.

  If you avoid plotting or images, you can have your widgets and your plots by building GDL
with "NO_WIDGET_DRAW" macro set.  If this is so, "X" or "WIN" plot drivers and widgets will coexist,
but widget_DRAW will return a widget ID of -1; the widget routine is prepared for this.
Under batch build conditions neither of these will be set.  OSX/Clang build fails a simple window
creation test when wx initialization occurs, and so the initialization, which is imperative for widget operation, is hidden in these two conditions.

  The UnsetFocus() is restored and plotting can be tested by running testsuite/testfocus
which will put up 5 plots and iconize two of them.
demo_widgets.pro and test_widgets.pro are both conditioned to expect a non-working widget_draw.