Added tests for QUERY_TIFF
These tests cover QUERY_TIFF on both tiled and untiled TIFF and GeoTIFF,
which returns structures with information about TIFF images. Currently
only files with a single image is tested, but the implementation should
support multiple images as well, and tests should be implemented.

The current implementation also supports READ_TIFF, but there are no
tests for this method yet. In order to fully test READ_TIFF, a variety
of sample images should be generated to test the most common settings.