Patch legacy HASH for /foldcase, tostruct(); file_link for WIN32
  The /recursive keyword for HASH::TOSTRUCT is implemented.  Also foldcase property
is now effective.  Test_hash is a bit more extensive.
  FILE_LINK built-in C++ function works in unix systems but for WIN32 the same library
procedure is not available; links are possible, however, so for the WIN32 builds the
internal function is disabled and a dummy procedure is placed in src/pro that returns
the same message the built-in would.

Greg Jung (aka maynardGK)
gvjung@gmail.com