Revise appveyor build for windows
   GDL is built in appveyor using mingw-w64 compiler and support libraries built with this
compiler.  scripts/appveyor_1.msys is where the files from the msys repository are obtained.
Some of those repository distros have been found to not be compatible with the appveyor build
and so "version upgrades" need to be done methodically.
  wxwidgets and graphicsmagick are the next two behemoth support libraries to be integrated.
The CMakelists.txt script is modified to survive the absence of these libraries, so that
the build command need not be changed to adjust the build.
Also, an entry for the plplot-driver selector USE_WINGDI_NO_WINGCC allow the wingdi.dll
device to be used - improvements in axis and labelling will be noticed, but there is
 still no anti-alias in the plot line.