Another failing experiment - prep wxWidgets for MINGW builds.
  I don't want to try to swap out gcc 7.4 which is used in MSYS/mingw32; this is
causing a problem however.  I can swap the mingw compiler path easier and so to prep for
a MINGW build from the respository distro, appveyor_plplot.msys moves files so that a windows-like
wxwidgets installation exists in win32libs/wxwidgets-3.0.4 which ca be found using MinGW Makefiles
option of Cmake.