Back out MAgick, Tiff, geotiff, add a few tests. insert scripts.
  scripts/mingw-bldgdl.bat and scripts/installgdl.ps1 are copied from the appveyor build.
  To replicate what appveyor does at home, make a C:\projects folder and make a symlink
called gdl pointing to the git. This can be done to a separate disk (with lots of space).
Also, make a symlink at top-level C: to mingw-w64 so that C:\mingw-w64\ points to where the
mingw-w64 distributions are kept (again, a different disk than C: which may have space constraints).
With these file system changes you can cut-and-paste to replicate the build done by appveyor.