Align structure data to intrinsic size (as for IDL)
This aligns the data stored in structures based on their intrinsic sizes.
Previously, use of EIGEN triggered storage on a uniform 16-byte boundary.  This gives no
preformance enhancement.  Yet it can greatly enlarge the physical memory size of structured
storage, which, in perverse cases, can result is reduced performance (due to greater memory spans required of the oprations).

Also included, to avoid an easy-to-make coding error, is conversion of structure field names
to upper case when DUStructDesc::AddTag() is invoked.