Return a value in GDLEventHandlerPy
The prototype of PyOS_InputHook requires to return a value (which is ignored,
however):

int (*PyOS_InputHook)(void);

Ignoring this and implementing a function without an explicit return seems to
cause gcc-9 to weirdly optimize away the `if( oldInputHook != NULL)`
condition, resulting in a segmentation fault in the next line (which then
tries to call the `NULL` `oldInputHook()`.

This patch just makes sure that the function always returns something.