include a posix option for windows' operation from a posix shell.
  Normally the file specification under windows is a mixed bag: the forward slash
is acceptable usage but the backslash is always present.  Under Unix, however, a backslash
disappears under normal shell operation - it is a special character in the posix shell.
For Instance, when running GDL from an MSYS2/cygwin posix shell, the user cannot use the
file reported by GDL without manually editing the string.
 For that reason much of the file_search output included the conversion to unix-style, (posix)
allowing inter-operation on common file systems, i.e. an NTFS disk can be mounted under
Linux and operated with as if it were another native-linux filesystem.  Nevertheless this
conflicts with standard windows' usage and IDL behavior.

 To allow for posix file output in GDL a "posixpaths" logical was created which can be set to
 "true" when run from windows

 1) by invoking the command-line argument "--posix" at startup.
 2)  A backdoor way

The system variable !GDL has been modified with a !GDL.GDL_POSIX variable to reflect this
 condition, it will be set =1 for any non-Win32 run and when run under Win32 with --posix.
Procedure PATH_SEP() uses this to provide a correct character- precautions have been made to
 side-step usage of !GDL.GDL_POSIX if it does not exist (an earlier GDL version, or IDL
  may be running the procedure).
If the backdoor way is used to set the logical, !GDL will remain unchanged.