added GDL_MAPS_DIR to Env. Var.
Unfortunately the default value for where lib etc... (and, thus, GDL_MAPS_DIR)  in the code is 'GDLDATADIR' which is defined in config.h as whatever location CMAKE thinks the installed GDL will go. As the build destination is not the same as the installer destination, the executable is not able to know where it is, unless we are more clever or we define manually the good value for all environment variables.