noticeable speedup of REBIN.
First attempt to patch inadequate parallelization code spread throughout all GDL:

Found that parallelization, using OPENMP and the 'if' clause, was not the best approach in many cases, and was definitely not right in all places:
- the if clause was wrong in two ways: 
 1) it would not test if tpool_nthread==1 , where it should just not parallelize, bang!
 2) the test on tpool_max_elts is wrong (probably a bad cut and paste that has been propagated everywhere)

Besides, experiences show that it is much faster to avoid entering a #pragma omp parallel section if parallelization is not needed at all.
This commit, and furher, should institue a new scheme:

 - parallelizing a loop must be done thus:
      bool parallelize= yes or no depending on various values, see below
      if (!parallelize) { //which is the most frequent
        for (...) // the loop
      } else { //jump in the same loop, but inside a parallel section:
#pragma omp parallel for num_threads(CpuTPOOL_NTHREADS) if (parallelize) (... eventually other necessary openmp clauses)
        for (...)
      }

The "if (parallelize)" above is unnecessary of course but may be handy in case the surrounding 'if (!parallelize) {} else {] ' disappears.

I do not understand why openmp+the if clause is blatantly slower than no openmp at all, but this is the repeated case on my machine, gcc+libgomp 8.4.0 

The minimal test for 'parallelize' is parallelize = (CpuTPOOL_NTHREADS > 1 && nE >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS >= nE));