closes #1643
Solve #1643 by forcing an update AFTER a refresh.
This because our wxwidget loop event is not in a separate thread and we need to activate it manually everytime something needs to be changed on the screen. And the loop needs to be convinced that a widget (and the plot is a widget) IS not up-to-date.
In contradistinction with X11 where the server loops independently from the program.